[BUILD][VMWARE] Keep i386 LLVM-MinGW guest frontier durable
Capture the current LLVM-MinGW i386 runtime, C++ compatibility, strict-build, loader, service, LSA, VMware guest-state, WDF, printing, and AP diagnostic work as a durable branch state before branch consolidation.

The strict Werror keep-going frontier is documented in the lab evidence; this commit preserves the owned source state for merge and further CCN reduction.

+++ sdk/lib/drivers/wdf/shared/irphandlers/pnp/fxpkgpnp.cpp
@@ -1691,7 +1691,8 @@ FxPkgPnp::QueryForCapabilities(
     if (NT_SUCCESS(status)) {
         ULONG states, i;
 
-        ASSERT(caps.DeviceCaps.DeviceWake <= 0xFF && caps.DeviceCaps.SystemWake <= 0xFF);
+        ASSERT((ULONG)caps.DeviceCaps.DeviceWake <= 0xFF &&
+               (ULONG)caps.DeviceCaps.SystemWake <= 0xFF);
 
         m_SystemWake = (BYTE) caps.DeviceCaps.SystemWake;
 
@@ -6491,4 +6492,3 @@ FxPkgPnp::PowerPolicyGetDeviceDeepestDeviceWakeState(
 
     return dxState;
 }
-
